@@ -1,15 +1,15 @@
 {
   "name": "react-native-power-translator",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "A no-hassle way to use Google and Microsoft cloud translation services in react native.",
   "main": "index.js",
   "scripts": {
-    "test": "jest --no-cache --watchAll",
-    "lint": "eslint --fix src/**/*.js"
+    "test": "node_modules/.bin/jest --no-cache --watchAll",
+    "lint": "./node_modules/.bin/eslint --fix src/**/*.js"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/danielkalbasi/react-native-power-translator.git"
+    "url": "git+https://github.com/danialkalbasi/react-native-power-translator.git"
   },
   "keywords": [
     "react-native",
@@ -23,12 +23,12 @@
     "google",
     "translate"
   ],
-  "author": "Daniel Kalbasi <danial.kalbassi@gmail.com> (www.danielkalbasi.com)",
+  "author": "Danial Kalbasi <danial.kalbassi@gmail.com> (www.danielkalbasi.com)",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/danielkalbasi/react-native-power-translator/issues"
+    "url": "https://github.com/danialkalbasi/react-native-power-translator/issues"
   },
-  "homepage": "https://github.com/danielkalbasi/react-native-power-translator#readme",
+  "homepage": "https://github.com/danialkalbasi/react-native-power-translator#readme",
   "jest": {
     "preset": "jest-expo",
     "collectCoverageFrom": [
@@ -43,24 +43,29 @@
     ]
   },
   "peerDependencies": {
-    "xmldom": "0.1.x"
+    "xmldom": "0.1.x",
+    "react": "^15.5.0"
   },
   "devDependencies": {
+    "babel-eslint": "^8.2.3",
     "babel-jest": "^20.0.3",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react-native-stage-0": "^1.0.1",
+    "eslint": "^4.19.1",
     "eslint-config-airbnb": "^15.0.1",
     "eslint-config-airbnb-base": "^12.0.0",
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-react": "^7.1.0",
-    "expo": "^21.0.0",
-    "jest-expo": "~21.0.0",
+    "expo": "^27.0.0",
+    "jest-cli": "^23.1.0",
+    "jest-expo": "~27.0.0",
     "lodash": "^4.17.4",
-    "react": "16.0.0-alpha.12",
-    "react-dom": "^16.0.0-beta.5",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-21.0.2.tar.gz",
-    "react-test-renderer": "16.0.0-alpha.12",
+    "prop-types": "^15.5.7",
+    "react": "16.3.1",
+    "react-dom": "^16.4.0",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-27.0.0.tar.gz",
+    "react-test-renderer": "16.4.0",
     "sinon": "4.0.1"
   }
 }
