@@ -1,6 +1,11 @@
 import PowerTranslator from './src/components/PowerTranslator';
 import ProviderTypes from './src/ProviderTypes';
-import Translation from './src/services';
+import { TranslatorFactory, TranslatorConfiguration } from './src/services';
 
 export default PowerTranslator;
-export { PowerTranslator, ProviderTypes, Translation }
+export {
+    PowerTranslator,
+    ProviderTypes,
+    TranslatorFactory,
+    TranslatorConfiguration
+};
